workers/requirements.txt
celery[redis]==5.3.0
boto3==1.28.0
minio==7.1.3
requests==2.31.0
Pillow==10.0.0
numpy==1.26.0
tqdm==4.66.1
ffmpeg-python==0.2.0
python-dotenv==1.0.0
qdrant-client==1.18.0
psutil==5.9.5
